split the reading and writing features
stegmp() now receives a file handler as second argument.

if this one is non-null, then it is the file from which we need to read
the input and write it back to the bitmap.

if it is null, then we need to read from the bitmap and display back the
data on the standard output

Signed-off-by: Olivier Gayot <duskcoder@gmail.com>